Move sensitive name providers to interface
